Added an errorHandler option.
Due to the quite special nature of JavaScript Promise, we can NOT simply pass the error in handler function to upper level (ANR user, normally http.createServer).

So either the user needs to handle the error in handler using try/catch block, or a universal errorHandler function must be provided to take the error.

In general: error can not break through the Promise level.